+++ HaloEzAPI/IHaloAPIService.cs
@@ -12,39 +12,40 @@ namespace HaloEzAPI
     public interface IHaloAPIService
     {
         //Stats
-        Task<PlayerMatches> GetMatchesForPlayer(string gamerTag, GameMode gameMode, int start, int count);
-        Task<ArenaPostGameReport> GetArenaPostGameCarnageReport(Guid matchId);
-        Task<CampaignPostGameReport> GetCampaignPostGameCarnageReport(Guid matchId);
-        Task<CustomPostGameReport> GetCustomPostGameCarnageReport(Guid matchId);
-        Task<WarzonePostGameReport> GetWarzonePostGameCarnageReport(Guid matchId);
-        Task<ArenaServiceRecordQueryResponse> GetArenaServiceRecords([MaxLength(32)] string[] players);
-        Task<CampaignServiceRecordQueryResponse> GetCampaignServiceRecords([MaxLength(32)] string[] players);
-        Task<CustomGameServiceRecordQueryResponse> GetCustomGameServiceRecords([MaxLength(32)] string[] players);
-        Task<WarzoneServiceRecordQueryResponse> GetWarzoneServiceRecords([MaxLength(32)] string[] players);
+        Task<PlayerMatches> GetMatchesForPlayer(string gamerTag, GameMode gameMode, int start, int count, bool bustCache = false);
+        Task<ArenaPostGameReport> GetArenaPostGameCarnageReport(Guid matchId, bool bustCache = false);
+        Task<CampaignPostGameReport> GetCampaignPostGameCarnageReport(Guid matchId, bool bustCache = false);
+        Task<CustomPostGameReport> GetCustomPostGameCarnageReport(Guid matchId, bool bustCache = false);
+        Task<WarzonePostGameReport> GetWarzonePostGameCarnageReport(Guid matchId, bool bustCache = false);
+        Task<ArenaServiceRecordQueryResponse> GetArenaServiceRecords([MaxLength(32)] string[] players, bool bustCache = false);
+        Task<CampaignServiceRecordQueryResponse> GetCampaignServiceRecords([MaxLength(32)] string[] players, bool bustCache = false);
+        Task<CustomGameServiceRecordQueryResponse> GetCustomGameServiceRecords([MaxLength(32)] string[] players, bool bustCache = false);
+        Task<WarzoneServiceRecordQueryResponse> GetWarzoneServiceRecords([MaxLength(32)] string[] players, bool bustCache = false);
         
         //MetaData
-        Task<IEnumerable<CampaignMission>> GetCampaignMissions();
-        Task<IEnumerable<Commendation>> GetCommendations();
-        Task<IEnumerable<Enemy>> GetEnemies();
-        Task<IEnumerable<FlexibleStat>> GetFlexibleStats();
-        Task<IEnumerable<GameBaseVariant>> GetGameBaseVariants();
-        Task<GameVariant> GetGameVariant(string id);
-        Task<IEnumerable<Impulse>> GetImpulses();
-        Task<MapVariant> GetMapVariant(string id);
-        Task<IEnumerable<Map>> GetMaps();
-        Task<IEnumerable<Medal>> GetMedals();
-        Task<IEnumerable<Playlist>> GetPlaylists();
-        Task<RequisitionPack> GetRequisitionPack(Guid id);
-        Task<RequisitionPack> GetRequisition(Guid id);
-        Task<IEnumerable<Skull>> GetSkulls();
-        Task<IEnumerable<SpartanRank>> GetSpartanRanks();
-        Task<IEnumerable<TeamColor>> GetTeamColours();
-        Task<IEnumerable<Vehicle>> GetVehicles();
-        Task<IEnumerable<Weapon>> GetWeapons();
-        Task<IEnumerable<Season>> GetSeasons();
+        Task<IEnumerable<CampaignMission>> GetCampaignMissions(bool bustCache = false);
+        Task<IEnumerable<Commendation>> GetCommendations(bool bustCache = false);
+        Task<IEnumerable<CSRDesignation>> GetCSRDesignations(bool bustCache = false);
+        Task<IEnumerable<Enemy>> GetEnemies(bool bustCache = false);
+        Task<IEnumerable<FlexibleStat>> GetFlexibleStats(bool bustCache = false);
+        Task<IEnumerable<GameBaseVariant>> GetGameBaseVariants(bool bustCache = false);
+        Task<GameVariant> GetGameVariant(string id, bool bustCache = false);
+        Task<IEnumerable<Impulse>> GetImpulses(bool bustCache = false);
+        Task<MapVariant> GetMapVariant(string id, bool bustCache = false);
+        Task<IEnumerable<Map>> GetMaps(bool bustCache = false);
+        Task<IEnumerable<Medal>> GetMedals(bool bustCache = false);
+        Task<IEnumerable<Playlist>> GetPlaylists(bool bustCache = false);
+        Task<RequisitionPack> GetRequisitionPack(Guid id, bool bustCache = false);
+        Task<RequisitionPack> GetRequisition(Guid id, bool bustCache = false);
+        Task<IEnumerable<Skull>> GetSkulls(bool bustCache = false);
+        Task<IEnumerable<SpartanRank>> GetSpartanRanks(bool bustCache = false);
+        Task<IEnumerable<TeamColor>> GetTeamColours(bool bustCache = false);
+        Task<IEnumerable<Vehicle>> GetVehicles(bool bustCache = false);
+        Task<IEnumerable<Weapon>> GetWeapons(bool bustCache = false);
+        Task<IEnumerable<Season>> GetSeasons(bool bustCache = false);
 
         //Profile
-        Task<Image> GetProfileEmblem(string gamerTag, int size = 256);
-        Task<Image> GetSpartanImage(string gamerTag, int size = 256, CropType cropType = CropType.Full);
+        Task<Image> GetProfileEmblem(string gamerTag, int size = 256, bool bustCache = false);
+        Task<Image> GetSpartanImage(string gamerTag, int size = 256, CropType cropType = CropType.Full, bool bustCache = false);
     }
 }